form css refactor + form buttons
