+++ styles/Form.module.css
@@ -4,17 +4,17 @@
     text-align: center;
 }
 
-.loginText {
+.formText {
     display: block;
 }
 
-.loginContainer {
+.formContainer {
     background-color: #4253b4;
     padding: 10px;
     display: inline-block;
 }
 
-.loginButtonContainer {
+.formButtonContainer {
     display: inline-block;
     padding: 4px;
     margin-top: 20px;
